refactor: fix magic strings implementation instead: - implement a new enum that represents the various mob spawn block kinds - implement this new enum in the places that need it to translate information correctly - defer to this implementation for the event handler
the only "magic" part of this has been shifted to the new MOB_SPAWN_BLOCKS
list inside DungeonMapBlocks -- if new instances of IGetMobSpawnBlockKind were implemented but
not added to this list, they would be missed for processing reasons.

Still seems a bit cleaner / better than the magic string version to me.